Bump template @types/react to 18.2.6 (#38713)
Summary:
Match the `react` dependency minor of `18.2` and specify a min version of `18.2.6`, which is the first version which includes `React.JSX` namespace. This enables usage of `React.JSX.Element`, now default as part of the template.

Note that `18.2.6` satisfied the previous semver, so new templates are already pulling in a version even newer than this, but updating projects may not do this automatically.

Changelog:
[General][Fixed] - Bump template types/react to 18.2.6

Pull Request resolved: https://github.com/facebook/react-native/pull/38713

Test Plan: CircleCI will build and typecheck a new template app.

Reviewed By: yungsters

Differential Revision: D47943524

Pulled By: NickGerleman

fbshipit-source-id: 49e587f8c2ebfce9fd1f00793a858d72ac0aa09e

Original-Commit: https://github.com/facebook/react-native/commit/b65dbf68d0b3172c36a499a07a0cbf98e8635acc