Make sure template is consuming the right buildToolsVersion - Take 2 (#40938)
Summary:
Currently, the template has a `buildToolsVersion = '34.0.0'` specified in the top level .gradle file but it's not currently using it.

This is causing the build to fallback to the default version provided by AGP which is 33.x
This is also causing the CI to download buildtools 34.0.0 as they're not in the container (causing network flakyness).

## Changelog:

[INTERNAL] [FIXED] - Make sure template is consuming the right buildToolsVersion

Pull Request resolved: https://github.com/facebook/react-native/pull/40938

Test Plan: CI should be green

Reviewed By: cipolleschi

Differential Revision: D50270482

Pulled By: cortinico

fbshipit-source-id: 09fdc66fe24f1cae760d07e4a2f044793a66cafc

Original: https://github.com/facebook/react-native/commit/a030224d1b69bf22df9c7bf757baeb10ebe89782