feat: create releases where PATCH doesn't have to match react-native (#43)
This feature needs to be released with react-native-community/cli#2475,
which decouples the 1:1 release model we have for templates and
react-native.

Switching users to this is going to be difficult, and should only be
rolled asap before 0.76. If we have to release a template fix, then only
users with the latest CLI will be able to use that and subsequent
versions.

I'm not sure of how to clearly communicate that to users.

- Added tests
- Set dry_run to default